requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
scikit-learn>=1.3
catboost>=1.2
shap>=0.50
matplotlib>=3.7
joblib>=1.3
jupyter>=1.0
nbformat>=5.0
tabulate>=0.9
